[refactor] modify MVC structure
