+++ .gitignore
@@ -6,6 +6,5 @@ node_modules
 
 # 練習用檔案
 upload
-test
-test1.jpg
+try
 images
