+++ app.js
@@ -4,27 +4,21 @@ import dotenv from "dotenv";
 dotenv.config();
 const { API_VERSION } = process.env;
 
-import connectToSocketIO from "./server.js";
+import connectToSocketIO from "./socketio.js";
 
 //創建 express 的物件
 const app = express();
 
-// 取得 static html file
-app.use(express.static("public"));
-
-// 可使用的 request body 格式
-app.use(express.json());
-
-// 取得 static html file
+// 取得 static file
 app.use(express.static("public"));
 
 // 可使用的 request body 格式
 app.use(express.json());
 
 // API routes
-import userRouter from "./routes/user_route.js";
-import chatRouter from "./routes/chat_record_route.js";
-import matchRouter from "./routes/match_route.js";
+import userRouter from "./server/routes/user_route.js";
+import chatRouter from "./server/routes/chat_record_route.js";
+import matchRouter from "./server/routes/match_route.js";
 app.use(`/api/${API_VERSION}`, [userRouter, chatRouter, matchRouter]);
 
 // 當使用者輸入錯的路徑，會直接掉進這個 middle ware
